+++ bittytax/parsers/__init__.py
@@ -1,5 +1,6 @@
 from . import binance
 from . import barclays
+from . import bitfinex
 from . import bitstamp
 from . import bittrex
 from . import changetip