+++ CHANGELOG.md
@@ -14,6 +14,6 @@ This is the initial beta release. Although it has been throughly tested, it's po
 - Exchange data files supported: Bitstamp, Bittrex, ChangeTip, Circle, Coinbase, Coinbase Pro, Coinfloor, Cryptopia, Cryptsy, Gatehub, OKEx, Poloniex, TradeSatoshi, Uphold.
 - Explorer data files supported: Etherscan.
 
-[Unreleased]: https://github.com/BittyTax/BittyTax/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/BittyTax/BittyTax/compare/v0.1.1...HEAD
 [0.1.1]: https://github.com/BittyTax/BittyTax/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/BittyTax/BittyTax/releases/tag/v0.1.0