+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Fixed
 - Fix for 'get_average_cost' exception when debug enabled.
 - Same-day buy pools should the use the timestamp of earliest transaction, not the latest, this prevents the possiblity of a negative balance.
+- Circle parser: added missing transaction types, 'internal_switch_currency' and 'switch_currency'.
 ### Added
 - Exchange data files: Wirex, Binance.