+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+
+## Version [0.1.2] Beta (2019-06-30)
 ### Fixed
 - Fix for 'get_average_cost' exception when debug enabled.
 - Same-day buy pools should the use the timestamp of earliest transaction, not the latest, this prevents the possiblity of a negative balance.
@@ -22,6 +24,7 @@ This is the initial beta release. Although it has been throughly tested, it's po
 - Exchange data files supported: Bitstamp, Bittrex, ChangeTip, Circle, Coinbase, Coinbase Pro, Coinfloor, Cryptopia, Cryptsy, Gatehub, OKEx, Poloniex, TradeSatoshi, Uphold.
 - Explorer data files supported: Etherscan.
 
-[Unreleased]: https://github.com/BittyTax/BittyTax/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/BittyTax/BittyTax/compare/v0.1.2...HEAD
+[0.1.2]: https://github.com/BittyTax/BittyTax/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/BittyTax/BittyTax/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/BittyTax/BittyTax/releases/tag/v0.1.0